Replace mpsc channel with mpmc channel in ThreadPool
This also removes the need for mutexes on the tx and rx fields. This is
mostly experimental and for my own testing, particularly since mpmc
channels have yet to be stablised.

I'm not sure why there was a mutex on the tx to begin with, so it's very
likely I've missed some important detail and this will harm performance
in my usecase instead of improving it like I'm hoping it will.

Potentially locking/parking perf of mutexes is preferable over std
channels, or something to do with poisoning of mutexes. Idk, if it's a
problem I'll probably find out pretty quickly.